# backend/api/OpenAPI.yml
openapi: 3.0.3
info:
  title: Banking Application API
  version: 2.5.0
  description: |
    REST API for a banking application with:
    - JWT authentication
    - customer registration and login
    - employee approval workflow
    - account management
    - separate transfer, deposit, and withdrawal creation endpoints
    - transaction history and filtering
    - short-lived JWT access tokens
    - refresh token rotation via secure HttpOnly cookie
    - JSON request and response bodies
    - pagination support for collection endpoints
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://bankapp.jxyden.dev/api
    description: Online host for the application and API
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Validation failed
        path:
          type: string
          example: /api/auth/register
...
    RegisterRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
        - phoneNumber
        - bsnNumber
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: SecurePassword123!
        phoneNumber:
          type: string
          example: "+31612345678"
        bsnNumber:
          type: string
          example: "123456789"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: SecurePassword123!
    AuthResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          example: 900
        user:
          $ref: '#/components/schemas/UserResponse'
    RefreshResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9...
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          example: 900
    UserResponse:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - phoneNumber
        - bsnNumber
        - role
        - approved
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          example: 1
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        phoneNumber:
          type: string
          example: "+31612345678"
        bsnNumber:
          type: string
          example: "123456789"
        role:
          type: string
          enum: [CUSTOMER, EMPLOYEE]
        approved:
          type: boolean
          example: false
        approvedAt:
          type: string
          format: date-time
          nullable: true
        approvedByUserId:
          type: integer
          format: int64
          nullable: true
          example: 10
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new customer
      description: |
        Registers a customer.
        The customer is created as not yet approved.
        After registration, the customer can log in but cannot use full banking features until approved by an employee.

        Authorization rules:
        - Public endpoint; no existing access token is required.
        - Registration always creates a CUSTOMER user.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Customer registered successfully
          headers:
            Location:
              description: URL of the created customer resource
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Customer already exists with same email or BSN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user and return access token
      description: |
        Authenticates the user and returns a short-lived access token.
        A refresh token is set by the server in a secure HttpOnly cookie.

        Authorization rules:
        - Public endpoint; no existing access token is required.
        - Available to existing CUSTOMER and EMPLOYEE users with valid credentials.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          headers:
            Set-Cookie:
              description: Secure HttpOnly refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: |
        Issues a new access token using the refresh token stored in a secure HttpOnly cookie.

        Authorization rules:
        - Public endpoint with respect to bearer authentication.
        - Requires a valid refresh token cookie issued by the server.
      security: []
      responses:
        '200':
          description: New access token issued successfully
          headers:
            Set-Cookie:
              description: Optional rotated refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Missing, invalid, or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...